minor changes, cleanups

+++ src/main/java/com/example/courier/dto/UserWithOrdersCountByStatus.java
@@ -2,5 +2,5 @@ package com.example.courier.dto;
 
 import com.example.courier.domain.User;
 
-public record UserWithOrdersCountByStatus(User user, int confirmedOrdersCount) {
+public record UserWithOrdersCountByStatus(User user, Long confirmedOrdersCount) {
 }
